[platform_device_loader] isolate load device related operations
