+++ libparsec/crates/platform_device_loader/src/device/load_ciphertext_key.rs
@@ -2,9 +2,39 @@
 
 use crate::{
     platform, AccountVaultOperationsFetchOpaqueKeyError, DeviceAccessStrategy,
-    LoadCiphertextKeyError, OpenBaoOperationsFetchOpaqueKeyError, RemoteOperationServer,
+    OpenBaoOperationsFetchOpaqueKeyError, RemoteOperationServer,
 };
 use libparsec_types::prelude::*;
+
+#[derive(Debug, thiserror::Error)]
+pub enum LoadCiphertextKeyError {
+    /// Typically returned if the access strategy doesn't match the device type
+    #[error("Invalid data")]
+    InvalidData,
+    #[error("Decryption failed")]
+    DecryptionFailed,
+    /// Note only a subset of load strategies requires server access to
+    /// fetch an opaque key that itself protects the ciphertext key
+    /// (e.g. account vault).
+    #[error("No response from {server} server: {error}")]
+    // We don't use `ConnectionError` here since this type only corresponds to
+    // an answer from the Parsec server and here any arbitrary server may have
+    // been (unsuccessfully) requested (e.g. OpenBao server).
+    RemoteOpaqueKeyFetchOffline {
+        server: RemoteOperationServer,
+        error: anyhow::Error,
+    },
+    /// Note only a subset of load strategies requires server access to
+    /// fetch an opaque key that itself protects the ciphertext key
+    /// (e.g. account vault).
+    #[error("{server} server opaque key fetch failed: {error}")]
+    RemoteOpaqueKeyFetchFailed {
+        server: RemoteOperationServer,
+        error: anyhow::Error,
+    },
+    #[error(transparent)]
+    Internal(anyhow::Error),
+}
 
 pub(super) async fn load_ciphertext_key(
     access: &DeviceAccessStrategy,
