+++ libparsec/crates/platform_device_loader/src/device/mod.rs
@@ -2,12 +2,16 @@
 
 mod archive;
 mod list;
+mod load;
+mod load_ciphertext_key;
 mod recovery;
 mod save;
 mod update;
 
 pub use archive::*;
 pub use list::*;
+pub use load::*;
+pub use load_ciphertext_key::*;
 pub use recovery::*;
 pub use save::*;
 pub use update::*;
